requirements.txt
Pillow>=10.0.0
Flask>=3.0.0
streamlit>=1.30.0
